pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "flai"
version = "8.0.0"
description = "Fully Local AI — a personal AI assistant that runs entirely on your hardware"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Valery Barval"},
]

dependencies = [
    "flask==2.3.3",
    "flask-wtf==1.2.1",
    "flask-limiter==3.5.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "gunicorn==21.2.0",
    "Pillow==10.1.0",
    "pytz==2023.3",
    "redis==5.0.1",
    "Flask-Babel==2.0.0",
    "qdrant-client==1.9.1",
    "PyPDF2==3.0.1",
    "python-docx==1.1.0",
    "python-magic==0.4.27",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.3",
    "pytest-flask==1.3.0",
    "pytest-cov==4.1.0",
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "types-requests",
    "types-redis",
    "types-PyPDF2",
    "types-python-docx",
    "types-pytz",
    "psycopg2-binary>=2.9.9",  # PostgreSQL support
]
postgres = [
    "psycopg2-binary>=2.9.9",
]
load-test = [
    "locust>=2.20.1",
]

[project.scripts]
flai = "wsgi:app"

[tool.setuptools.packages.find]
include = ["app*", "modules*"]

[tool.ruff]
target-version = "py39"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "translations",
]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "PTH",   # flake8-use-pathlib
]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # do not perform function calls in argument defaults (common in Flask)
    "PTH123", # open() should be replaced by Path.open() (too invasive)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__ are expected

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
exclude = ["tests/", "venv/", "translations/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (with mocked services)",
    "e2e: End-to-end tests (requires running services)",
    "slow: Slow running tests (>1 second)",
    "requires_db: Tests that require database access",
    "requires_redis: Tests that require Redis access",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::UserWarning:flask_limiter",
    "ignore::UserWarning:qdrant_client",
]
